Fix SML version of raytracer benchmark.
The SML version of the raytracer benchmark was mistakenly generating
random spheres with radii drawn from (0.0,1.0), rather than with radii
drawn from (0.01,0.08) as in the PML version.

This revealed a potential Overflow exception in the conversion of a
(real * real * real) color to a (Word8.word * Word8.word * Word8.word)
color.  The real color components should always be in the range
(0.0,1.0), but numerical artifacts can cause it to exceed 1.0 and the
conversion Real.round(x * 255.0) can overflow.  Introduce a "clampd"
function as is present in the image.c library of the PML runtime.